Move server test utils to a common location
